+++ frontend/src/components/PageWrapper/PageWrapper.tsx
@@ -2,16 +2,21 @@
 // © Crown Copyright 2025. This work has been developed by the National Digital Twin Programme
 // and is legally attributed to the Department for Business and Trade (UK) as the governing entity.
 
-import { ReactNode } from 'react';
+import { PropsWithChildren } from 'react';
 import { Box } from '@mui/material';
 
-const PageWrapper = ({ children }: { children: ReactNode }) => (
+type PageWrapperProps = PropsWithChildren & {
+  backgroundColor?: string;
+};
+
+const PageWrapper = ({ children, backgroundColor = 'transparent' }: PageWrapperProps) => (
   <Box
     display="flex"
     flexDirection="column"
     gap={4}
     paddingY="2rem"
     sx={{
+      backgroundColor,
       '@media print': {
         padding: 0,
         marginTop: '1rem'
